cluster adopts machines orphan
if a machine is not created with a machineset then the cluster will
attempt to adopt it. machines no longer add owner references for cluster
or secrets. the cluster secret is also does not have an owner reference
so it must be deleted when the cluster is deleted. Cluster finalizer
has been removed so the owned objects will be deleted. This removes the
need for a "StoppingClusterPhase". Most of that logic is moved to
cluster not found section.

Signed-off-by: Andrew O'Neill <foolusion@gmail.com>
Change-Id: I850b7bb5fef966c6630d38202ac5d306eb0d4b6e